+++ src/ksignal.c
@@ -193,7 +193,7 @@ void sigqueue_init(sigqueue_t * list, struct proc * p)
  * @param type Syscall type
  * @param p Syscall parameters
  */
-uint32_t ksignal_syscall(uint32_t type, void * p)
+uintptr_t ksignal_syscall(uint32_t type, void * p)
 {
     switch(type) {
         case SYSCALL_SIGNAL_KILL:
